+++ src/tests/cache-control-tests.ts
@@ -155,6 +155,7 @@ describe('cache-control tests', () => {
       miner_txid: '0x4321',
       canonical: true,
       hash: '0x1234',
+      index_block_hash: '0xdeadbeef',
       height: 1,
       parent_block_hash: '0xff0011',
       parent_microblock_hash: '',
@@ -298,6 +299,7 @@ describe('cache-control tests', () => {
       miner_txid: '0x4321',
       canonical: true,
       hash: '0x1234',
+      index_block_hash: '0xdeadbeef',
       height: 1,
       parent_block_hash: '0xff0011',
       parent_microblock_hash: '',
@@ -423,6 +425,156 @@ describe('cache-control tests', () => {
     expect(request8.status).toBe(200);
     expect(request8.type).toBe('application/json');
     expect(request8.headers['etag']).toBeUndefined();
+  });
+
+  test('transaction cache control', async () => {
+    const txId1 = '0x0153a41ed24a0e1d32f66ea98338df09f942571ca66359e28bdca79ccd0305cf';
+    const txId2 = '0xfb4bfc274007825dfd2d8f6c3f429407016779e9954775f82129108282d4c4ce';
+
+    const block1 = new TestBlockBuilder({
+      block_height: 1,
+      index_block_hash: '0x01',
+    })
+      .addTx()
+      .build();
+    await db.update(block1);
+
+    // No tx yet.
+    const request1 = await supertest(api.server).get(`/extended/v1/tx/${txId1}`);
+    expect(request1.status).toBe(404);
+    expect(request1.type).toBe('application/json');
+
+    // Add mempool tx.
+    const mempoolTx1 = testMempoolTx({ tx_id: txId1 });
+    await db.updateMempoolTxs({ mempoolTxs: [mempoolTx1] });
+
+    // Valid mempool ETag.
+    const request2 = await supertest(api.server).get(`/extended/v1/tx/${txId1}`);
+    expect(request2.status).toBe(200);
+    expect(request2.type).toBe('application/json');
+    expect(request2.headers['etag']).toBeTruthy();
+    const etag1 = request2.headers['etag'];
+
+    // Cache works with valid ETag.
+    const request3 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag1);
+    expect(request3.status).toBe(304);
+    expect(request3.text).toBe('');
+
+    // Mine the same tx into a block
+    const block2 = new TestBlockBuilder({
+      block_height: 2,
+      index_block_hash: '0x02',
+      parent_index_block_hash: '0x01',
+    })
+      .addTx({ tx_id: txId1 })
+      .build();
+    await db.update(block2);
+
+    // Cache no longer works with mempool ETag but we get updated ETag.
+    const request4 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag1);
+    expect(request4.status).toBe(200);
+    expect(request4.headers['etag']).toBeTruthy();
+    const etag2 = request4.headers['etag'];
+
+    // Cache works with new ETag.
+    const request5 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag2);
+    expect(request5.status).toBe(304);
+    expect(request5.text).toBe('');
+
+    // No tx #2 yet.
+    const request6 = await supertest(api.server).get(`/extended/v1/tx/${txId2}`);
+    expect(request6.status).toBe(404);
+    expect(request6.type).toBe('application/json');
+
+    // Tx #2 directly into a block
+    const block3 = new TestBlockBuilder({
+      block_height: 3,
+      index_block_hash: '0x03',
+      parent_index_block_hash: '0x02',
+    })
+      .addTx({ tx_id: txId2 })
+      .build();
+    await db.update(block3);
+
+    // Valid block ETag.
+    const request7 = await supertest(api.server).get(`/extended/v1/tx/${txId2}`);
+    expect(request7.status).toBe(200);
+    expect(request7.type).toBe('application/json');
+    expect(request7.headers['etag']).toBeTruthy();
+    const etag3 = request7.headers['etag'];
+
+    // Cache works with valid ETag.
+    const request8 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId2}`)
+      .set('If-None-Match', etag3);
+    expect(request8.status).toBe(304);
+    expect(request8.text).toBe('');
+
+    // Oops, new blocks came, all txs before are non-canonical
+    const block2a = new TestBlockBuilder({
+      block_height: 2,
+      index_block_hash: '0x02ff',
+      parent_index_block_hash: '0x01',
+    })
+      .addTx({ tx_id: '0x1111' })
+      .build();
+    await db.update(block2a);
+    const block3a = new TestBlockBuilder({
+      block_height: 3,
+      index_block_hash: '0x03ff',
+      parent_index_block_hash: '0x02ff',
+    })
+      .addTx({ tx_id: '0x1112' })
+      .build();
+    await db.update(block3a);
+    const block4 = new TestBlockBuilder({
+      block_height: 4,
+      index_block_hash: '0x04',
+      parent_index_block_hash: '0x03ff',
+    })
+      .addTx({ tx_id: '0x1113' })
+      .build();
+    await db.update(block4);
+
+    // Cache no longer works for tx #1.
+    const request9 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag2);
+    expect(request9.status).toBe(200);
+    expect(request9.headers['etag']).toBeTruthy();
+    const etag4 = request9.headers['etag'];
+
+    // Cache works again with new ETag.
+    const request10 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag4);
+    expect(request10.status).toBe(304);
+    expect(request10.text).toBe('');
+
+    // Mine tx in a new block
+    const block5 = new TestBlockBuilder({
+      block_height: 5,
+      index_block_hash: '0x05',
+      parent_index_block_hash: '0x04',
+    })
+      .addTx({ tx_id: txId1 })
+      .build();
+    await db.update(block5);
+
+    // Make sure old cache for confirmed tx doesn't work, because the index_block_hash has changed.
+    const request11 = await supertest(api.server)
+      .get(`/extended/v1/tx/${txId1}`)
+      .set('If-None-Match', etag2);
+    expect(request11.status).toBe(200);
+    expect(request11.headers['etag']).toBeTruthy();
+    const etag5 = request11.headers['etag'];
+    expect(etag2).not.toBe(etag5);
   });
 
   afterEach(async () => {